+++ js/adapters.ts
@@ -27,6 +27,8 @@ export const leafletRasterLayer = (source: PMTiles, options: any) => {
             mimeType = "image/jpeg";
           } else if (header.tileType == 4) {
             mimeType = "image/webp";
+          } else if (header.tileType == 5) {
+            mimeType = "image/avif";
           }
         });
         loaded = true;
@@ -121,16 +123,21 @@ export class Protocol {
         this.tiles.set(pmtiles_url, instance);
       }
 
-      // TODO: create vector_layers if present to return valid TileJSON
+      instance
+        .getHeader()
+        .then((h) => {
+          const tilejson = {
+            tiles: [params.url + "/{z}/{x}/{y}"],
+            minzoom: h.minZoom,
+            maxzoom: h.maxZoom,
+            bounds: [h.minLon, h.minLat, h.maxLon, h.maxLat],
+          };
+          callback(null, tilejson, null, null);
+        })
+        .catch((e) => {
+          callback(e, null, null, null);
+        });
 
-      instance.getHeader().then((h) => {
-        const tilejson = {
-          tiles: [params.url + "/{z}/{x}/{y}"],
-          minzoom: h.minZoom,
-          maxzoom: h.maxZoom,
-        };
-        callback(null, tilejson, null, null);
-      });
       return {
         cancel: () => {},
       };
@@ -181,7 +188,7 @@ export class Protocol {
           })
           .catch((e) => {
             if ((e as Error).name !== "AbortError") {
-              throw e;
+              callback(e, null, null, null);
             }
           });
       });