+++ js/index.ts
@@ -155,14 +155,22 @@ type DecompressFunc = (
   compression: Compression
 ) => Promise<ArrayBuffer>;
 
-async function fflateDecompress(
+async function defaultDecompress(
   buf: ArrayBuffer,
   compression: Compression
 ): Promise<ArrayBuffer> {
   if (compression === Compression.None || compression === Compression.Unknown) {
     return buf;
   } else if (compression === Compression.Gzip) {
-    return decompressSync(new Uint8Array(buf));
+    if (typeof (globalThis as any).DecompressionStream == "undefined") {
+      return decompressSync(new Uint8Array(buf));
+    } else {
+      let stream = new Response(buf).body!;
+      let result: ReadableStream<Uint8Array> = stream.pipeThrough(
+        new (globalThis as any).DecompressionStream("gzip")
+      );
+      return new Response(result).arrayBuffer();
+    }
   } else {
     throw Error("Compression method not supported");
   }
@@ -174,6 +182,7 @@ export enum TileType {
   Png = 2,
   Jpeg = 3,
   Webp = 4,
+  Avif = 5,
 }
 
 const HEADER_SIZE_BYTES = 127;
@@ -273,13 +282,19 @@ export class FileAPISource implements Source {
 
 export class FetchSource implements Source {
   url: string;
+  customHeaders: Headers;
 
-  constructor(url: string) {
+  constructor(url: string, customHeaders: Headers = new Headers()) {
     this.url = url;
+    this.customHeaders = customHeaders;
   }
 
   getKey() {
     return this.url;
+  }
+
+  setHeaders(customHeaders: Headers) {
+    this.customHeaders = customHeaders;
   }
 
   async getBytes(
@@ -294,9 +309,15 @@ export class FetchSource implements Source {
       signal = controller.signal;
     }
 
+    const requestHeaders = new Headers(this.customHeaders);
+    requestHeaders.set(
+      "Range",
+      "bytes=" + offset + "-" + (offset + length - 1)
+    );
+
     let resp = await fetch(this.url, {
       signal: signal,
-      headers: { Range: "bytes=" + offset + "-" + (offset + length - 1) },
+      headers: requestHeaders,
     });
 
     // TODO: can return 416 with offset > 0 if content changed, which will have a blank etag.
@@ -547,7 +568,7 @@ export class ResolvedValueCache {
   constructor(
     maxCacheEntries = 100,
     prefetch = true,
-    decompress: DecompressFunc = fflateDecompress
+    decompress: DecompressFunc = defaultDecompress
   ) {
     this.cache = new Map<string, ResolvedValue>();
     this.maxCacheEntries = maxCacheEntries;
@@ -682,7 +703,7 @@ export class SharedPromiseCache {
   constructor(
     maxCacheEntries = 100,
     prefetch = true,
-    decompress: DecompressFunc = fflateDecompress
+    decompress: DecompressFunc = defaultDecompress
   ) {
     this.cache = new Map<string, SharedPromiseCacheValue>();
     this.maxCacheEntries = maxCacheEntries;
@@ -824,7 +845,7 @@ export class PMTiles {
     if (decompress) {
       this.decompress = decompress;
     } else {
-      this.decompress = fflateDecompress;
+      this.decompress = defaultDecompress;
     }
     if (cache) {
       this.cache = cache;